WIP: Move CacheCtrl
