+++ podpac/core/cache/__init__.py
@@ -3,3 +3,4 @@ from podpac.core.cache.cache_ctrl import CacheCtrl, get_default_cache_ctrl, make
 from podpac.core.cache.ram_cache_store import RamCacheStore
 from podpac.core.cache.disk_cache_store import DiskCacheStore
 from podpac.core.cache.s3_cache_store import S3CacheStore
+from podpac.core.cache.cache import NoCacheMixin, DiskCacheMixin
